Remove Python interface and consolidate functionality in cffi.buildtool CLI
